accelerator-utilities/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "accelerator-utilities"
version = "0.1.0"
description = "Framework-agnostic utilities, no dependencies on other accelerator packages. Utility functions and common helpers for the accelerator framework"
readme = "README.md"
requires-python = ">=3.9"
authors = [
    {name = "Compression & Acceleration Team"},
    {name = "Ryabykin Alexey r00926208", email=""}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
]
dependencies = [
    "torch>=2.5.1",
    "hydra-core>=1.3.2",
    "omegaconf>=2.3.0",
    "rich",
    "onnx",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "black",
    "isort",
    "flake8",
    "pre-commit",
]

[tool.setuptools.packages.find]
where = ["src"]
include = ["accelerator*"]

[tool.setuptools.package-dir]
"" = "src"

[tool.black]
line-length = 120
target-version = ['py39']

[tool.isort]
profile = "black"
line_length = 120
